Fix: 위도 경도 저장 타입 변경

+++ backend/product-service/src/main/java/com/example/productservice/dto/response/ReservationResponseDto.java
@@ -25,9 +25,9 @@ public class ReservationResponseDto {
 
     private Integer price;
 
-    private Float lng;
+    private Double lng;
 
-    private Float lat;
+    private Double lat;
 
     private String location;
